# 2015-2016-api-mobiliteitscentrum/swagger.json
{
    "swagger": "2.0",
    "info": {
        "title": "Mobiliteitscentrum Gent Website API (VOPRO 2015-1016)",
        "description": "API specificatie voor Milestone 3.\n",
        "version": "3.0.0"
    },
    "host": "voproX.ugent.be",
    "schemes": [
        "http"
    ],
    "basePath": "/api",
    "produces": [
        "application/json"
    ],
    "paths": {
...
        "/user/forgot_password": {
            "post": {
                "summary": "Endpoint for when a user forgets his/her password. The backend only has 1 response, 200 OK, since we don't want to return the (non-)existance of a user. The backend will send an email if we find a user which matches this email. If we don't, we do nothing and return 200 OK.\n",
                "parameters": [
                    {
                        "name": "email",
                        "in": "body",
                        "description": "The email linked to this user.",
                        "required": true,
                        "schema": {
                            "type": "object",
                            "properties": {
                                "email": {
                                    "type": "string",
                                    "description": "The email of the user."
                                }
                            }
                        }
                    },
                    {
                        "name": "Authorization",
                        "in": "header",
                        "description": "A bearer to perform captcha validation",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Always returned, even if the email is not linked to any user."
                    }
                }
            }
        },
        "/user/{user_id}/": {
            "get": {
                "summary": "Returns the user",
                "parameters": [
                    {
                        "name": "user_id",
                        "in": "path",
                        "description": "The unique identifier of this user",
                        "required": true,
                        "type": "number"
                    },
                    {
                        "name": "Authorization",
                        "in": "header",
                        "description": "The access token as a bearer",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "The user",
                        "schema": {
                            "$ref": "#/definitions/User"
                        }
                    },
                    "401": {
                        "description": "The user is not logged in",
                        "schema": {
                            "$ref": "#/definitions/Error"
                        }
                    },
                    "403": {
                        "description": "The requested user is not the current user or the current user is not an Administrator",
                        "schema": {
                            "$ref": "#/definitions/Error"
                        }
                    },
                    "404": {
                        "description": "User does not exist",
                        "schema": {
                            "$ref": "#/definitions/Error"
                        }
                    }
                }
            },
            "put": {
                "summary": "Edits the user",
                "parameters": [
                    {
                        "name": "user_id",
                        "in": "path",
                        "description": "The unique identifier of this user",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "user",
                        "in": "body",
                        "description": "User input parameters",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/User"
                        }
                    },
                    {
                        "name": "Authorization",
                        "in": "header",
                        "description": "The access token as a bearer",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "The modified user",
                        "schema": {
                            "$ref": "#/definitions/User"
                        }
                    },
                    "401": {
                        "description": "The user is not logged in",
                        "schema": {
                            "$ref": "#/definitions/Error"
                        }
                    },
                    "403": {
                        "description": "The requested user is not the current user or the current user is not an Administrator",
                        "schema": {
                            "$ref": "#/definitions/Error"
                        }
                    },
                    "404": {
                        "description": "User does not exist",
                        "schema": {
                            "$ref": "#/definitions/Error"
                        }
                    },
                    "409": {
                        "description": "The user already exists (ex. email already in use)",
                        "schema": {
                            "$ref": "#/definitions/Error"
                        }
                    }
                }
            },
            "delete": {
                "summary": "Deletes the user",
                "parameters": [
                    {
                        "name": "user_id",
                        "in": "path",
                        "description": "The unique identifier of this user",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "Authorization",
                        "in": "header",
                        "description": "The access token as a bearer",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "204": {
                        "description": "The user was deleted"
                    },
                    "401": {
                        "description": "The user is not logged in",
                        "schema": {
                            "$ref": "#/definitions/Error"
                        }
                    },
                    "403": {
                        "description": "The requested user is not the current user or the current user is not an Administrator",
                        "schema": {
                            "$ref": "#/definitions/Error"
                        }
                    },
                    "404": {
                        "description": "User does not exist",
                        "schema": {
                            "$ref": "#/definitions/Error"
                        }
                    }
                }
            }
        },
...
        "/user/{user_id}/travel/{travel_id}/route/": {
            "get": {
                "summary": "Returns all routes for this travel",
                "parameters": [
                    {
                        "name": "user_id",
                        "in": "path",
                        "description": "The unique identifier of this user",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "travel_id",
                        "in": "path",
                        "description": "The unique identifier of this travel",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "Authorization",
                        "in": "header",
                        "description": "The access token as a bearer",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "All routes of this travel",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/Route"
                            }
                        }
                    },
                    "401": {
                        "description": "The user is not logged in",
                        "schema": {
                            "$ref": "#/definitions/Error"
                        }
                    },
                    "403": {
                        "description": "The requested user is not the current user or the current user is not an Administrator",
                        "schema": {
                            "$ref": "#/definitions/Error"
                        }
                    },
                    "404": {
                        "description": "User or travel does not exist",
                        "schema": {
                            "$ref": "#/definitions/Error"
                        }
                    }
                }
            },
            "post": {
                "summary": "Creates a new route on this travel",
                "parameters": [
                    {
                        "name": "user_id",
                        "in": "path",
                        "description": "The unique identifier of this user",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "travel_id",
                        "in": "path",
                        "description": "The unique identifier of this travel",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "route",
                        "in": "body",
                        "description": "The new route's parameters",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/Route"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "The created route",
                        "schema": {
                            "$ref": "#/definitions/Route"
                        }
                    },
                    "401": {
                        "description": "The user is not logged in",
                        "schema": {
                            "$ref": "#/definitions/Error"
                        }
                    },
                    "403": {
                        "description": "The requested user is not the current user or the current user is not an Administrator",
                        "schema": {
                            "$ref": "#/definitions/Error"
                        }
                    },
                    "404": {
                        "description": "User or travel does not exist",
                        "schema": {
                            "$ref": "#/definitions/Error"
                        }
                    },
                    "409": {
                        "description": "The route already exists",
                        "schema": {
                            "$ref": "#/definitions/Error"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "User": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string",
                    "description": "Unique identifier representing a specific User. This attribute is not required in a request, but is required in a response."
                },
                "links": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/Link"
                    }
                },
                "password": {
                    "type": "string",
                    "description": "User password in clear text. Please save this in a secure way. Clearly, this attribute may not be returned in a response.\n"
                },
                "email": {
                    "type": "string",
                    "description": "Email address of the user."
                },
                "cell_number": {
                    "type": "string",
                    "description": "Cellphone number of the user."
                },
                "mute_notifications": {
                    "type": "boolean",
                    "description": "Master switch to disable all notifications (ex. when on holiday)"
                },
                "validated": {
                    "type": "object",
                    "description": "Indicates which fields the user has validated through external services",
                    "properties": {
                        "email": {
                            "type": "boolean",
                            "description": "Indicates if this user has validated his/her email"
                        },
                        "cell_number": {
                            "type": "boolean",
                            "description": "Indicates if this user has validated his/her cell number"
                        }
                    }
                }
            },
            "required": [
                "password",
                "email",
                "mute_notifications"
            ]
        },
...
        "Coordinates": {
            "type": "object",
            "properties": {
                "lat": {
                    "type": "number",
                    "format": "double",
                    "description": "Latitude"
                },
                "lon": {
                    "type": "number",
                    "format": "double",
                    "description": "Longitude"
                }
            },
            "required": [
                "lat",
                "lon"
            ]
        },
...
        "Route": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string",
                    "description": "Unique identifier for  this route"
                },
                "waypoints": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/Coordinates"
                    }
                },
                "transportation_type": {
                    "type": "string",
                    "description": "Indicates which type of transportion this route uses (ex. bus, car, bike)"
                },
                "notify_for_event_types": {
                    "type": "array",
                    "description": "User has to mark specific EventTypes as relevant  for this route to receive notifications for events of these types",
                    "items": {
                        "$ref": "#/definitions/EventType"
                    }
                },
                "notify": {
                    "type": "object",
                    "description": "The mediums on which the user wishes to be notified of events on this route",
                    "properties": {
                        "email": {
                            "type": "boolean",
                            "description": "User wishes to be notified via email"
                        },
                        "cell_number": {
                            "type": "boolean",
                            "description": "User wishes to be notified via SMS"
                        }
                    }
                },
                "active": {
                    "type": "boolean",
                    "description": "Boolean to indicate if the user wants to receive notifications for this route"
                }
            },
            "required": [
                "waypoints",
                "transportation_type",
                "notify_for_event_types",
                "notify",
                "active"
            ]
        },
...
        "EventType": {
            "type": "object",
            "properties": {
                "type": {
                    "type": "string",
                    "description": "Type of event"
                }
            }
        },
...
        "Link": {
            "type": "object",
            "properties": {
                "rel": {
                    "type": "string",
                    "description": "Indicates the relation of the linked object (ex. self)"
                },
                "href": {
                    "type": "string",
                    "description": "The actual link in full URL (ex. https://vopro.ugent.be/api/user/123/)"
                }
            }
        },
        "Error": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer",
                    "format": "int32",
                    "description": "HTTP status code (ex. 400)"
                },
                "message": {
                    "type": "string",
                    "description": "Human readable error (ex. You are not logged in)"
                },
                "fields": {
                    "type": "string",
                    "description": "Missing or incorrect fields for input"
                }
            }
        },
...
    }
}